chore(apps/api): refactor
