+++ apps/api/src/services/specs/getFuseLeveragedTokensByChainId.spec.ts
@@ -1,4 +1,4 @@
-import { getFuseLeveragedTokensByChainId } from "./flts";
+import { getFuseLeveragedTokensByChainId } from "../flts";
 import { ChainId } from "@risedle/types";
 
 describe("getFuseLeveragedTokensByChainId", () => {
