+++ apps/api/src/services/specs/getTokenInfoFromPancakeSwapSubgraph.spec.ts
@@ -1,4 +1,4 @@
-import { getTokenInfoFromPancakeSwapSubgraph } from "./tokens";
+import { getTokenInfoFromPancakeSwapSubgraph } from "../tokens";
 
 describe("getTokenInfoFromPancakeSwapSubgraph", () => {
     describe("given empty array", () => {
